% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_approve.R
\name{ApprovedPostHTML}
\alias{ApprovedPostHTML}
\title{Create HTML report - approved post - to be sent to the mailing list}
\usage{
ApprovedPostHTML(dt_post)
}
\arguments{
\item{dt_post}{A tibble element that contains all relevant information about the post}
}
\description{
Create HTML report - approved post - to be sent to the mailing list
}
