% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_approval_fct_aux.R
\name{ConfirmationPostHTML}
\alias{ConfirmationPostHTML}
\title{HTML file to be sent to the poster confirming that
his/her post was approved.}
\usage{
ConfirmationPostHTML(id_request)
}
\arguments{
\item{id_request}{string, id of the request}
}
\description{
HTML file to be sent to the poster confirming that
his/her post was approved.
}
