% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_panel.R
\name{f7Post}
\alias{f7Post}
\title{Specific function for the UI of the offers' panel}
\usage{
f7Post(..., subject = NULL, date = NULL, footer = NULL)
}
\arguments{
\item{subject}{description of the offer}

\item{date}{date when the offer was posted}

\item{footer}{Footer content, if any. Must be wrapped in a tagList}
}
\description{
Specific function for the UI of the offers' panel
}
