% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_approval_fct_aux.R
\name{gm_attach_url}
\alias{gm_attach_url}
\title{Auxiliary function to attach files from URL to message - gmailr package}
\usage{
gm_attach_url(mime, description)
}
\arguments{
\item{mime}{message}

\item{description}{complete URL including scheme 
(such as http://, https://, ftp:// or file://)}
}
\description{
Auxiliary function to attach files from URL to message - gmailr package
}
