% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_subscribe.R
\name{is_valid_email}
\alias{is_valid_email}
\title{Checks the validity of an email input}
\usage{
is_valid_email(x)
}
\arguments{
\item{x}{string, email input}
}
\description{
Checks the validity of an email input
}
