% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_post_fct_aux.R
\name{PostApprovalHTML}
\alias{PostApprovalHTML}
\title{Create HTML report to be sent to the post approvers}
\usage{
PostApprovalHTML(
  id_post,
  name_poster,
  email_poster,
  type_post = c("jobs", "services", "upcycle"),
  subject,
  description,
  contact_email,
  contact_phone
)
}
\arguments{
\item{id_post}{string, name of the poster}

\item{name_poster}{string, name of the poster}

\item{email_poster}{string, email of the poster}

\item{type_post}{string, type of post. It can be "jobs" or "services"}

\item{subject}{string, subject of the post}

\item{description}{string, description of the post}

\item{contact_email}{string, email to contact about the post}

\item{contact_phone}{string, phone number to contact about the post}
}
\description{
Create HTML report to be sent to the post approvers
}
