% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_approval_fct_aux.R
\name{RejectionPostHTML}
\alias{RejectionPostHTML}
\title{HTML file to be sent to the poster stating that
his/her post was rejected.}
\usage{
RejectionPostHTML(id_request, comment)
}
\arguments{
\item{id_request}{string, id of the request}

\item{comment}{string, comments related to the post rejection}
}
\description{
HTML file to be sent to the poster stating that
his/her post was rejected.
}
