% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_post_fct_aux.R
\name{UploadPost}
\alias{UploadPost}
\title{Upload post data to Googlesheets}
\usage{
UploadPost(
  name_poster,
  email_poster,
  type_post = c("jobs", "services", "upcycle"),
  subject,
  description,
  contact_email,
  contact_phone,
  files_tmp,
  gdpr_acceptance
)
}
\arguments{
\item{name_poster}{string, name of the poster}

\item{email_poster}{string, email of the poster}

\item{type_post}{string, type of post. It can be "jobs", "services" or "upcycle"}

\item{subject}{string, subject of the post}

\item{description}{string, description of the post}

\item{contact_email}{string, email to contact about the post}

\item{contact_phone}{string, phone number to contact about the post}

\item{files_tmp}{dataframe that contains the names, sizes, MIME types and datapaths of the uploaded files}

\item{gdpr_acceptance}{boolean, whether the poster has accepted or not the}
}
\description{
Upload post data to Googlesheets
}
