% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_approval_fct_aux.R
\name{ApprovePost}
\alias{ApprovePost}
\title{Send approval confirmation mail to poster and send approved post
to the mailing list}
\usage{
ApprovePost(id_request, id_approver, comment)
}
\arguments{
\item{id_request}{string, id of the request}

\item{id_approver}{string, id of the post approver}

\item{comment}{string, comments related to the post approval}
}
\description{
Send approval confirmation mail to poster and send approved post
to the mailing list
}
