% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_panel.R, R/mod_share_fct_aux.R
\name{f7Post}
\alias{f7Post}
\title{Specific function for the UI of the offers' panel}
\usage{
f7Post(
  subject,
  date,
  type = c("jobs", "services", "upcycle"),
  content,
  footer = NULL
)

f7Post(
  subject,
  date,
  type = c("jobs", "services", "upcycle"),
  content,
  footer = NULL
)
}
\arguments{
\item{subject}{post brief description}

\item{date}{date of the post}

\item{type}{type of post, it can take any of the following:
c('jobs', 'services', 'upcycle')}

\item{content}{post description and details}

\item{footer}{Footer content, if any. Must be wrapped in a tagList}
}
\description{
Specific function for the UI of the offers' panel

Function to create UI posts given inputs
}
