% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_share_fct_aux.R
\name{f7Post}
\alias{f7Post}
\title{Function to create UI posts given inputs}
\usage{
f7Post(subject, date, content, footer = NULL)
}
\arguments{
\item{subject}{post brief description}

\item{date}{date of the post}

\item{content}{post description and details}

\item{footer}{Footer content, if any. Must be wrapped in a tagList}
}
\description{
Function to create UI posts given inputs
}
