% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_approval_fct_aux.R
\name{RejectPost}
\alias{RejectPost}
\title{Sends rejection mail to the poster with the corresponding explanation}
\usage{
RejectPost(id_request, id_approver, comment, session)
}
\arguments{
\item{id_request}{string, id of the request}

\item{id_approver}{string, id of the post approver}

\item{comment}{string, comments related to the post rejection}

\item{session}{shiny session}
}
\description{
Sends rejection mail to the poster with the corresponding explanation
}
